# services/staffing-service/openapi.yaml
openapi: 3.1.0
info:
  title: CareCareer Staffing Service
  version: 0.1.0
  description: |
    Manages facilities, departments, credential requirements, and workers.
    All endpoints require Bearer token authentication (RS256).
    All operations are tenant-scoped via validated session principal.
servers:
  - url: http://localhost:3200
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/audit:
    get:
      operationId: listAuditEntries
      summary: List audit entries for tenant
      parameters:
        - name: action
          in: query
          schema: { type: string }
        - name: resourceType
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Audit entries
  /v1/audit/{resourceType}/{resourceId}:
    get:
      operationId: getResourceAudit
      summary: Get audit history for a specific resource
      parameters:
        - name: resourceType
          in: path
          required: true
          schema: { type: string }
        - name: resourceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Resource audit history
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed platform access token
...